Add 1st Project of DAN Explore_Weather_Trends .
Explore_Weather_Trends: (Data Analyst Nanodegree)
Extract the data from the database.
There's a workspace in the previous section that is connected to a database.
You’ll need to export the temperature data for the world as well as for the closest big city to where you live. 
You can find a list of cities and countries in the city_list table. 
To interact with the database, you'll need to write a SQL query.
Write a SQL query to extract the city level data. Export to CSV.
Write a SQL query to extract the global data. Export to CSV.
Open up the CSV in whatever tool you feel most comfortable using. 
We suggest using Excel or Google sheets, but you are welcome to use another tool, such as Python or R.
Create a line chart that compares your city’s temperatures with the global temperatures. 
Make sure to plot the moving average rather than the yearly averages in order to smooth out the lines, making trends more observable.
Make observations about the similarities and differences between the world averages and your city’s averages, as well as overall trends.

Review:
The SQL query used to extract the data is included.
The query runs without error and pulls the intended data.
Moving averages are calculated to be used in the line chart.
A line chart is included in the submission.
The chart and its axes have titles, and there's a clear legend.
The student includes four observations about their provided data visualization.
The four observations are accurate.